+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [UCX-3764] Allow user to hide the first listed room
 * [UCX-3769] Fix incoming message does not scroll window down
 * [UCX-3766] Move the flex-nav scollbars to the right of page
+* [UCX-3731] Fix rendering create channel after more channels
 
 ## 0.2.3 (2018-02-02)